feat(sputnik): support for custom build version
